Changed to symbol collector approach.
