+++ Src/Sentry.Xamarin/SentryXamarin.cs
@@ -37,10 +37,11 @@ namespace Sentry
             options.RegisterNativeActivityStatus();
             options.RegisterNativeIntegrations();
             options.RegisterXamarinEventProcessors();
-            options.RegisterScreenshotEventProcessor();
             options.RegisterXamarinInAppExclude();
             options.ProtocolPackageName ??= ProtocolPackageName;
             SentrySdk.Init(options);
+
+            options.RegisterScreenshotEventProcessor();
         }
     }
 }
